requirements.txt
numpy
pandas
matplotlib
seaborn
ipykernel
scikit-learn
xgboost
shap
Flask
joblib
google-generativeai